Adicionando melhorias de código

+++ MyFinance.Business/Service/IAccountService.cs
@@ -8,8 +8,8 @@ public interface IAccountService
     Task<bool> ExistsAsync(int id);
     Task<Account> GetByIdAsync(int value);
     Task<List<Account>> GetListAsync();
-    Task<Account> AddAsync(string name, double balance);
-    Task UpdateAsync(int id, string name, double balance);
+    Task<Account> AddAsync(string name, decimal balance);
+    Task UpdateAsync(int id, string name, decimal balance);
     Task DeleteAsync(int id);
 
 }
